Move SymbolDeclarator interface to JP core module
